+++ language/dutch_utf-8.php
@@ -405,6 +405,8 @@ $LANG_EVLIST = array(
 'dates'         => 'Dates',
 'shortcode'     => 'Short Code',
 'free_caps'     => 'FREE',
+'click_for_details' => 'Click for details.',
+'logo_image'    => 'Logo Image',
 );
 
 $LANG_EVLIST_HELP = array(